o modified node and relationship hashCode implementations, they   now just return the id o fixed some int multiplications in node/relationship/property   store implementations that resulted in values > 2^31 (caused   everything from illegal position to wrong records accessed) o modified persistence window pool not to allocate memory   mapped windows at startup, instead first "refresh bricks" will   allocate (better when we have huge stores). o "refresh bricks" and "expand bricks" implementation in   persistence window pool have been tweaked to perform better   for very large stores
git-svn-id: https://svn.neo4j.org/projects/neo/trunk@241 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ enterprise/src/java/org/neo4j/impl/nioneo/store/PropertyIndexStore.java
@@ -127,8 +127,8 @@ public class PropertyIndexStore extends AbstractStore implements Store
 		if ( buffer != null && !hasWindow( id ) )
 		{
 			buffer.makeReadyForTransfer();
-			getFileChannel().transferTo( id * RECORD_SIZE, RECORD_SIZE, 
-				buffer.getFileChannel() );
+			getFileChannel().transferTo( ((long) id) * RECORD_SIZE, 
+				RECORD_SIZE, buffer.getFileChannel() );
 			ByteBuffer buf = buffer.getByteBuffer();
 			byte inUse = buf.get();
 			assert inUse == Record.IN_USE.byteValue();
@@ -166,8 +166,8 @@ public class PropertyIndexStore extends AbstractStore implements Store
 		if ( buffer != null && !hasWindow( id ) )
 		{
 			buffer.makeReadyForTransfer();
-			getFileChannel().transferTo( id * RECORD_SIZE, RECORD_SIZE, 
-				buffer.getFileChannel() );
+			getFileChannel().transferTo( ((long) id) * RECORD_SIZE, 
+				RECORD_SIZE, buffer.getFileChannel() );
 			ByteBuffer buf = buffer.getByteBuffer();
 			byte inUse = buf.get();
 			assert inUse == Record.IN_USE.byteValue();
@@ -235,7 +235,7 @@ public class PropertyIndexStore extends AbstractStore implements Store
 	private PropertyIndexRecord getRecord( int id, Buffer buffer ) 
 		throws IOException
 	{
-		int offset = ( id - buffer.position() ) * getRecordSize();
+		int offset = (int) ( id - buffer.position() ) * getRecordSize();
 		buffer.setOffset( offset );
 		boolean inUse = ( buffer.get() == Record.IN_USE.byteValue() );
 		if ( !inUse )
@@ -267,7 +267,7 @@ public class PropertyIndexStore extends AbstractStore implements Store
 		throws IOException
 	{
 		int id = record.getId();
-		int offset = ( id - buffer.position() ) * getRecordSize();
+		int offset = (int) ( id - buffer.position() ) * getRecordSize();
 		buffer.setOffset( offset );
 		if ( record.inUse() )
 		{
